moving create_icon/create_icon in gui.utils.utils with create_font
