+++ packages/pymodaq_gui/src/pymodaq_gui/managers/manager_base.py
@@ -6,7 +6,7 @@ from qtpy.QtCore import Qt, QModelIndex
 from qtpy import QtWidgets, QtCore, QtGui
 from qtpy.QtGui import QKeySequence
 
-from pymodaq_gui.managers.action_manager import create_icon
+from pymodaq_gui.utils.utils import create_icon
 
 from pymodaq_utils.logger import set_logger, get_module_name
 from pymodaq.extensions.utils import CustomExt
