+++ packages/pymodaq_gui/src/pymodaq_gui/utils/widgets/push.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from qtpy import QtWidgets, QtCore, QtGui
 from qtpy.QtWidgets import QWidget
 
-from pymodaq_gui.managers.action_manager import create_icon
+from pymodaq_gui.utils.utils import create_icon
 from pymodaq_gui.utils.widgets.spinbox import SpinBox
